+++ CUIElement.h
@@ -104,7 +104,7 @@ public:
     /// </summary>
     void set_config(stream_reader * reader, size_t size, abort_callback & abortHandler) final
     {
-        _MainState.Read(reader, size, abortHandler);
+        _UIThread.Read(reader, size, abortHandler);
     }
 
     /// <summary>
@@ -112,7 +112,7 @@ public:
     /// </summary>
     void get_config(stream_writer * writer, abort_callback & abortHandler) const final
     {
-        _MainState.Write(writer, abortHandler);
+        _UIThread.Write(writer, abortHandler);
     }
 
     #pragma endregion