+++ Configuration/ConfigurationDialog.cpp
@@ -1,5 +1,5 @@
 
-/** $VER: ConfigurationDialog.cpp (2025.09.18) P. Stuer - Implements the configuration dialog. **/
+/** $VER: ConfigurationDialog.cpp (2025.09.22) P. Stuer - Implements the configuration dialog. **/
 
 #include "pch.h"
 #include "ConfigurationDialog.h"
@@ -193,12 +193,15 @@ BOOL ConfigurationDialog::OnInitDialog(CWindow w, LPARAM lParam)
             { IDC_COLOR_ORDER, L"Determines how to sort the colors selected from the artwork." },
 
             { IDC_ARTWORK_BACKGROUND, L"Displays artwork on the graph background." },
+            { IDC_ARTWORK_TYPE, L"Specifies which artwork will be show on the graph background." },
 
             { IDC_FIT_MODE, L"Determines how over- and undersized artwork is rendered." },
             { IDC_FIT_WINDOW, L"Use the component window size instead of the client area of the graph to fit the artwork." },
 
             { IDC_ARTWORK_OPACITY, L"Determines the opacity of the artwork when displayed." },
             { IDC_FILE_PATH, L"A fully-qualified file path or a foobar2000 script that returns the file path of an image to display on the graph background" },
+
+            { IDC_LOG_LEVEL, L"Sets the verbosity of the log information that gets written to the console." },
 
             // Graphs
             { IDC_GRAPH_SETTINGS, L"Shows the list of graphs." },
@@ -918,7 +921,35 @@ void ConfigurationDialog::Initialize()
         w.SetCurSel((int) _State->_ColorOrder);
     }
     {
+        auto w = (CComboBox) GetDlgItem(IDC_ARTWORK_TYPE);
+
+        w.ResetContent();
+
+        for (const auto & x : { L"Front", L"Back", L"Disc", L"Icon", L"Artist" })
+            w.AddString(x);
+
+        w.SetCurSel((int) _State->_ArtworkType);
+    }
+    {
         GetDlgItem(IDC_FILE_PATH).SetWindowTextW(_State->_ArtworkFilePath.c_str());
+    }
+    #pragma endregion
+
+    #pragma region Component
+    {
+        auto w = (CComboBox) GetDlgItem(IDC_LOG_LEVEL);
+
+        w.ResetContent();
+
+        int i = -1;
+
+        for (const auto & Text : { L"Never", L"Fatal", L"Error", L"Warn", L"Info", L"Debug", L"Trace", L"Always", })
+        {
+            w.AddString(Text);
+
+            if (++i == CfgLogLevel)
+                w.SetCurSel((int) i);
+        }
     }
     #pragma endregion
 
@@ -1427,11 +1458,25 @@ void ConfigurationDialog::OnSelectionChanged(UINT notificationCode, int id, CWin
             break;
         }
 
+        case IDC_ARTWORK_TYPE:
+        {
+            _State->_ArtworkType = (ArtworkType) SelectedIndex;
+            break;
+        }
+
         case IDC_FIT_MODE:
         {
             _State->_FitMode = (FitMode) SelectedIndex;
 
             UpdateCommonPage();
+            break;
+        }
+
+        case IDC_LOG_LEVEL:
+        {
+            CfgLogLevel.set((int64_t) SelectedIndex);
+
+            Log.SetLevel((LogLevel) CfgLogLevel.get());
             break;
         }
 
@@ -2661,6 +2706,7 @@ void ConfigurationDialog::OnButtonClick(UINT, int id, CWindow)
         case IDC_RESET:
         {
             _State->Reset();
+            CfgLogLevel.set((int64_t) DefaultCfgLogLevel);
 
             Initialize();
             break;
@@ -3033,14 +3079,20 @@ void ConfigurationDialog::UpdatePages(size_t index) const noexcept
             IDC_SMOOTHING_METHOD, IDC_SMOOTHING_METHOD_LBL, IDC_SMOOTHING_FACTOR, IDC_SMOOTHING_FACTOR_LBL,
             IDC_SHOW_TOOLTIPS, IDC_SUPPRESS_MIRROR_IMAGE,
 
+        // Artwork
         IDC_ARTWORK,
             IDC_ARTWORK_BACKGROUND,
+            IDC_ARTWORK_TYPE_LBL, IDC_ARTWORK_TYPE,
             IDC_FIT_MODE_LBL, IDC_FIT_MODE, IDC_FIT_WINDOW,
             IDC_ARTWORK_OPACITY_LBL, IDC_ARTWORK_OPACITY, IDC_ARTWORK_OPACITY_SPIN, IDC_ARTWORK_OPACITY_LBL_2,
             IDC_FILE_PATH_LBL, IDC_FILE_PATH,
             IDC_NUM_ARTWORK_COLORS_LBL, IDC_NUM_ARTWORK_COLORS, IDC_NUM_ARTWORK_COLORS_SPIN,
             IDC_LIGHTNESS_THRESHOLD_LBL, IDC_LIGHTNESS_THRESHOLD, IDC_LIGHTNESS_THRESHOLD_SPIN, IDC_LIGHTNESS_THRESHOLD_LBL_2,
             IDC_COLOR_ORDER_LBL, IDC_COLOR_ORDER,
+
+        // Component
+        IDC_COMPONENT,
+            IDC_LOG_LEVEL_LBL, IDC_LOG_LEVEL
     };
 
     static const int Page5[] =