+++ Configuration/Layout.h
@@ -1,5 +1,5 @@
 
-/** $VER: Layout.h (2024.08.18) P. Stuer - Defines the layout of the configuration dialog. **/
+/** $VER: Layout.h (2025.09.22) P. Stuer - Defines the layout of the configuration dialog. **/
 
 #pragma once
 
@@ -824,13 +824,25 @@
     #define X_G02    X_B06 + 5
     #define Y_G02    Y_G12 + H_G12 + IY
 
+    // Label: Artwork Type
+    #define W_G40    66
+    #define H_G40    H_LBL
+    #define X_G40    X_G08
+    #define Y_G40    Y_G02 + H_G02 + IY
+
+    // Combobox:  Artwork Type
+    #define W_G41    86
+    #define H_G41    H_CBX
+    #define X_G41    X_G40 + W_G40 + IX
+    #define Y_G41    Y_G40
+
     #pragma region Fit mode
 
     // Label: Fit mode
     #define W_G30    66
     #define H_G30    H_LBL
     #define X_G30    X_G02
-    #define Y_G30    Y_G02 + H_G02 + IY
+    #define Y_G30    Y_G41 + H_G41 + IY
 
     // Combobox: Fit mode
     #define W_G31    86
@@ -881,7 +893,27 @@
     #pragma endregion
 
 #define W_B06  184
-#define H_B06   11 + H_G07 + IY + H_G09 + IY + H_G12 + IY + H_G02 + IY + H_G31 + IY + H_G32 + IY + H_G04 + IY + H_G14 + 7
+#define H_B06   11 + H_G07 + IY + H_G09 + IY + H_G12 + IY + H_G02 + IY + H_G41 + IY + H_G31 + IY + H_G32 + IY + H_G04 + IY + H_G14 + 7
+
+#pragma region Component
+// Groupbox
+#define X_K00   X_B06
+#define Y_K00   Y_B06 + H_B06 + IY
+#define W_K00   W_B06
+
+    // Label: Log level
+    #define X_K01    X_K00 +  5
+    #define Y_K01    Y_K00 + 11
+    #define W_K01    34
+    #define H_K01    H_LBL
+
+    // Combobox: Log level
+    #define X_K02    X_K01 + W_K01 + IX
+    #define Y_K02    Y_K01
+    #define W_K02    48
+    #define H_K02    H_CBX
+
+#define H_K00   11 + H_K02 + 7
 
 #pragma endregion