+++ UIElement.h
@@ -19,7 +19,7 @@
 /// <summary>
 /// Implements the UIElement and Playback interface.
 /// </summary>
-class uielement_t : public CWindowImpl<uielement_t>, private play_callback_impl_base
+class uielement_t : public CWindowImpl<uielement_t>, private play_callback_impl_base, now_playing_album_art_notify
 {
 public:
     uielement_t();
@@ -109,6 +109,8 @@ private:
     void DeleteTrackingToolTip() noexcept;
     graph_t * GetGraph(const CPoint & pt) noexcept;
 
+    void on_album_art(album_art_data::ptr data);
+
     void GetAlbumArtFromTrack(const metadb_handle_ptr & track, abort_callback & abort);
     void GetAlbumArtFromScript(const metadb_handle_ptr & track, abort_callback & abort);
 
@@ -163,8 +165,8 @@ private:
     #pragma endregion
 
 protected:
-    state_t _MainState;
-    state_t _ThreadState;
+    state_t _UIThread;
+    state_t _RenderThread;
 
     msc::critical_section_t _CriticalSection;
     ConfigurationDialog _ConfigurationDialog;