+++ docs/History.md
@@ -1,6 +1,17 @@
 
 # foo_vis_spectrum_analyzer History
 
+v0.8.0.0-beta3, 2025-09-19
+
+* New: [User guide](docs/README.md) documenting all the configuration settings.
+* New: When the `Tool tips` option is disabled you can still get a tool tip when you click and hold the left mousebutton over a frequency band.
+* Improved: Radial Bars
+  * Implemented horizontal and amplitude-based gradient, peak mode.
+* Fixed: Weird behavior in IIR filter bank/analog-style analyzer. (Provided by TF3RDL)
+* Fixed: Stability of the tooltip.
+* Fixed: SWIFT analyzer (Regression)
+* Fixed: Artwork color sorting (Regression)
+
 v0.8.0.0-beta2, 2024-08-18
 
 * New: Radial Bar visualization.